Replace Disqus with Discord
Removes DISQUS integration if favour of hinting users to join the Discord. The comment field per-page is not used at all and only makes the website look unprofessional (especially the "What did you think of this? React with an emoji!"). And even **if** someone used it then it would need to be moderated, and I'm sure much better support would possible to give on Discord instead.