Resolve design problems with new theme on mobile
Fix top header with text wrapping on thin mobile devices.

Made tables scrolleable on horizontal axis for when table doesn't fit. Ideally many pages should be rewritten to avoid table, as it is very difficult to make mobile-compatible, but this is waaay outside the scope of this PR.

Made the text in the Discord-footer-box scale depending on the screen width.